moving footer into a separate folder
